+++ src/factor_instances_provider/agnostic_paths/mod.rs
@@ -1,7 +1,9 @@
 mod derivation_preset;
 mod index_agnostic_path;
+mod quantified_derivation_preset;
 mod quantities;
 
 pub use derivation_preset::*;
 pub use index_agnostic_path::*;
+pub use quantified_derivation_preset::*;
 pub use quantities::*;
