+++ src/factor_instances_provider/agnostic_paths/quantified_derivation_preset.rs
@@ -1,0 +1,17 @@
+use crate::prelude::*;
+
+#[derive(Clone, Copy, Hash, PartialEq, Eq, derive_more::Debug)]
+#[debug("🎯: {:?} #{}", self.derivation_preset, self.quantity)]
+pub struct QuantifiedDerivationPreset {
+    pub derivation_preset: DerivationPreset,
+    pub quantity: usize,
+}
+
+impl QuantifiedDerivationPreset {
+    pub fn new(derivation_preset: DerivationPreset, quantity: usize) -> Self {
+        Self {
+            derivation_preset,
+            quantity,
+        }
+    }
+}
