+++ src/factor_instances_provider/agnostic_paths/quantities.rs
@@ -1,3 +1,4 @@
 use crate::prelude::*;
 
+/// The quantity of DerivationPreset's to fill cache with.
 pub const CACHE_FILLING_QUANTITY: usize = 30;
