+++ src/factor_instances_provider/next_index_assigner/next_derivation_entity_index_cache_analyzing_assigner.rs
@@ -8,7 +8,7 @@ impl NextDerivationEntityIndexCacheAnalyzingAssigner {
         Self { cache }
     }
 
-    pub fn next(
+    fn max(
         &self,
         factor_source_id: FactorSourceIDFromHash,
         index_agnostic_path: IndexAgnosticPath,
@@ -16,6 +16,23 @@ impl NextDerivationEntityIndexCacheAnalyzingAssigner {
         let max = self
             .cache
             .max_index_for(factor_source_id, index_agnostic_path);
+        Ok(max)
+    }
+
+    /// Returns the next index for the given `FactorSourceIDFromHash` and
+    /// `IndexAgnosticPath`, by analyzing the cache. In case of read failure
+    /// will this method return `Err`, if the cache did not contain any data for
+    /// the given `FactorSourceIDFromHash` and `IndexAgnosticPath`, then `Ok(None)` is returned.
+    ///
+    /// If some index was found, this method returns `max + 1`.
+    ///
+    /// Can also fail if addition of one would overflow.
+    pub fn next(
+        &self,
+        factor_source_id: FactorSourceIDFromHash,
+        index_agnostic_path: IndexAgnosticPath,
+    ) -> Result<Option<HDPathComponent>> {
+        let max = self.max(factor_source_id, index_agnostic_path)?;
         let Some(max) = max else { return Ok(None) };
         max.add_one().map(Some)
     }
