+++ src/factor_instances_provider/next_index_assigner/next_derivation_entity_index_with_ephemeral_offsets_for_factor_source.rs
@@ -6,13 +6,28 @@ use crate::prelude::*;
 /// on top of next index analysis based on cache or profile. This is used so that
 /// the FactorInstanceProvider can consecutively call `next` N times to get a range of
 /// of `N` unique indices, added to the otherwise next based on cache/profile analysis.
-#[derive(Debug, Default)]
+#[derive(Debug)]
 pub struct NextDerivationEntityIndexWithEphemeralOffsetsForFactorSource {
+    factor_source_id: FactorSourceIDFromHash,
     ephemeral_offsets: RwLock<HashMap<IndexAgnosticPath, HDPathValue>>,
 }
 
 impl NextDerivationEntityIndexWithEphemeralOffsetsForFactorSource {
-    pub fn reserve(&self, agnostic_path: IndexAgnosticPath) -> HDPathValue {
+    pub fn empty(factor_source_id: FactorSourceIDFromHash) -> Self {
+        Self {
+            factor_source_id,
+            ephemeral_offsets: RwLock::new(HashMap::new()),
+        }
+    }
+
+    /// Returns the next free index for the FactorSourceID and IndexAgnosticPath,
+    /// and increases the local ephemeral offset.
+    pub fn reserve(
+        &self,
+        factor_source_id: FactorSourceIDFromHash,
+        agnostic_path: IndexAgnosticPath,
+    ) -> HDPathValue {
+        assert_eq!(self.factor_source_id, factor_source_id);
         let mut binding = self.ephemeral_offsets.write().unwrap();
         if let Some(existing) = binding.get_mut(&agnostic_path) {
             let free = *existing;
